Updating MPI IO results

+++ LearningPythonMPI/MPI_IO/GenMatrices.py
@@ -6,9 +6,9 @@ mat_size = int(sys.argv[1])
 size = int(sys.argv[2])
 iteration = int(sys.argv[3])
 
-mat_A = np.random.rand(mat_size,mat_size)
+mat_A = np.random.rand(mat_size,mat_size).astype(np.float32)
 #As B is an arbitrary matrix, we will take the transpose of it
-mat_B = np.random.rand(mat_size,mat_size)
+mat_B = np.random.rand(mat_size,mat_size).astype(np.float32)
 mat_B = np.transpose(mat_B)
 
 np.savetxt(f"mat_A/mat_A_{mat_size}_{size}_{iteration}.txt", mat_A)
